Linux Support Added
Added a few lines to make the serial work for linux. Graphing, writing
files and everthing else has been tested. Check for permission errors
before using on new computers, either run script as sudo (not good),
or chmod a+rw the serial port. Will look for ways to do this in script